requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
Django==4.2.1
django-colorfield==0.8.0
django-cors-headers==4.0.0
django-js-asset==2.0.0
django-mptt==0.14.0
django-mptt-admin==2.5.0
djangorestframework==3.14.0
Pillow==9.5.0
psycopg2==2.9.6
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.6.2
tzdata==2023.3
